Worked on Unholy Assault
Unholy Assault is mostly working. When played from the hand it is
dropped on top of a creature. If a creature card is underneath it when
it is dropped it is placed in an open summon slot and begins counting
down. Once the timer is done it applied the debuff to the creature and
is sent to the graveyard. If no summon zone is available then it is
automatically sent to the graveyard..

-Added a few get/set functions in Card.cs
-Added creatureDebuffCardIsDropped fuction in the player controller.